Remove process tag
We now model everything that process tags handled in Cocina.

Fixes #4045
