+++ spec/requests/create_object_spec.rb
@@ -952,85 +952,4 @@ RSpec.describe 'Create object' do
       end
     end
   end
-
-  context 'when type does not have a process tag (e.g., webarchive-binary)' do
-    before do
-      allow(SolrService).to receive(:get).and_return(no_result)
-      allow(AdministrativeTags).to receive(:create)
-    end
-
-    let(:expected) do
-      build(:dro, id: 'druid:gg777gg7777', type: Cocina::Models::ObjectType.webarchive_binary, label: 'This is my label', title: 'This is my label', admin_policy_id: admin_policy_id).new(
-        identification: { sourceId: 'warc:999999' }
-      )
-    end
-    let(:data) do
-      <<~JSON
-        {#{' '}
-          "cocinaVersion":"#{Cocina::Models::VERSION}",
-          "type":"#{Cocina::Models::ObjectType.webarchive_binary}",
-          "label":"This is my label","version":1,"access":{},
-          "administrative":{"hasAdminPolicy":"#{admin_policy_id}"},
-          "identification":{"sourceId":"warc:999999"},
-          "structural":{}}
-      JSON
-    end
-    let(:access_template) do
-      {
-        view: 'dark',
-        download: 'none'
-      }
-    end
-
-    it 'registers the object and does not create process tag' do
-      post '/v1/objects',
-           params: data,
-           headers: { 'Authorization' => "Bearer #{jwt}", 'Content-Type' => 'application/json' }
-      expect(response.body).to equal_cocina_model(expected)
-      expect(response.status).to eq(201)
-      expect(response.location).to eq '/v1/objects/druid:gg777gg7777'
-      expect(AdministrativeTags).not_to have_received(:create)
-    end
-  end
-
-  context 'when type does have a process tag' do
-    before do
-      allow(SolrService).to receive(:get).and_return(no_result)
-      allow(AdministrativeTags).to receive(:create)
-    end
-
-    let(:expected) do
-      build(:dro, id: 'druid:gg777gg7777', label: 'This is my label', title: 'This is my label', admin_policy_id: admin_policy_id).new(
-        identification: { sourceId: 'warc:999999' }
-      )
-    end
-    let(:data) do
-      <<~JSON
-        {#{' '}
-          "cocinaVersion":"#{Cocina::Models::VERSION}",
-          "type":"#{Cocina::Models::ObjectType.object}",
-          "label":"This is my label","version":1,"access":{},
-          "administrative":{"hasAdminPolicy":"#{admin_policy_id}"},
-          "identification":{"sourceId":"warc:999999"},
-          "structural":{}}
-      JSON
-    end
-    let(:access_template) do
-      {
-        view: 'dark',
-        download: 'none'
-      }
-    end
-
-    it 'registers the object and creates process tag' do
-      post '/v1/objects',
-           params: data,
-           headers: { 'Authorization' => "Bearer #{jwt}", 'Content-Type' => 'application/json' }
-      expect(response.body).to equal_cocina_model(expected)
-      expect(response.status).to eq(201)
-      expect(response.location).to eq '/v1/objects/druid:gg777gg7777'
-      expect(AdministrativeTags).to have_received(:create).with(identifier: 'druid:gg777gg7777',
-                                                                tags: ['Process : Content Type : File'])
-    end
-  end
 end
