+++ spec/services/cocina_object_store_spec.rb
@@ -223,7 +223,7 @@ RSpec.describe CocinaObjectStore do
       allow(Cocina::ObjectValidator).to receive(:validate)
       allow(Notifications::ObjectCreated).to receive(:publish)
       allow(store).to receive(:merge_access_for).and_return(requested_cocina_object)
-      allow(store).to receive(:add_tags_for_create)
+      allow(store).to receive(:add_project_tag)
       allow(SuriService).to receive(:mint_id).and_return(druid)
       allow(SynchronousIndexer).to receive(:reindex_remotely_from_cocina)
       allow(EventFactory).to receive(:create)
@@ -241,7 +241,7 @@ RSpec.describe CocinaObjectStore do
         expect(Notifications::ObjectCreated).to have_received(:publish)
         expect(Cocina::ObjectValidator).to have_received(:validate).with(requested_cocina_object)
         expect(store).to have_received(:merge_access_for).with(requested_cocina_object)
-        expect(store).to have_received(:add_tags_for_create).with(druid, requested_cocina_object)
+        expect(store).to have_received(:add_project_tag).with(druid, requested_cocina_object)
         expect(SynchronousIndexer).to have_received(:reindex_remotely_from_cocina)
         expect(ObjectVersion.current_version(druid).tag).to eq('1.0.0')
         expect(EventFactory).to have_received(:create).with(druid: druid, event_type: 'registration', data: Hash)
@@ -259,7 +259,7 @@ RSpec.describe CocinaObjectStore do
         expect(Notifications::ObjectCreated).to have_received(:publish)
         expect(Cocina::ObjectValidator).to have_received(:validate).with(requested_cocina_object)
         expect(store).to have_received(:merge_access_for).with(requested_cocina_object)
-        expect(store).to have_received(:add_tags_for_create).with(druid, requested_cocina_object)
+        expect(store).to have_received(:add_project_tag).with(druid, requested_cocina_object)
         expect(SynchronousIndexer).to have_received(:reindex_remotely_from_cocina)
         expect(ObjectVersion.current_version(druid).tag).to eq('1.0.0')
         expect(EventFactory).to have_received(:create).with(druid: druid, event_type: 'registration', data: Hash)
